+++ public/ckeditor/plugins/articleTeaser/lang/en.js
@@ -1,0 +1,8 @@
+'use strict'
+
+;(function (CKEDITOR) {
+    CKEDITOR.plugins.setLang('articleTeaser', 'en', {
+        title: 'Article teaser',
+        idRequired: 'Please define article id (number)',
+    })
+})(CKEDITOR)